# src/main/java/dk/systemedz/entsoe/marketdataservice/domain/models/PricePoint.java
package dk.systemedz.entsoe.marketdataservice.domain.models;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.LocalDateTime;

@AllArgsConstructor
@NoArgsConstructor
@Data
@Builder
public class PricePoint {
    private LocalDateTime hour;
    private BigDecimal price;
}