# src/main/resources/openapi/entsoe-price-service-api-v1.yml
openapi: 3.0.2
info:
  title: ENTSO-E - Price Service
  version: 1.0.0
  description: >-
    ##  ENTSO-E Service
    
    
    This service is made with ther purpose of simplifying the current API that is provided by  ENTSO-E,
    and also converting the responses into JSON format instead of the current XML format provided by
    ENTSO-E.
    
    
    This API has been developed for Grønnere Elforbrug, by Kenneth Thomsen.
  contact:
    name: Kenneth Thomsen
    url: 'https://elforbrug.nu'
    email: kontakt@kthomsen.dk
  license:
    name: GNU GPLv3
    url: 'https://www.gnu.org/licenses/gpl.txt'
servers:
  -
    url: 'https://entsoe.systemedz.com/api/{version}'
    description: Production environment - currently hosted by Systemedz (Kenneth Thomsen)
    variables:
      version:
        enum:
          - v1
        default: v1
        description: The API Version
  -
    url: 'https://localhost:8443/api/{version}'
    description: Localhost Environment (for testing)
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
        description: The API Version
paths:
...
  '/prices/{areaCode}/by-interval/day/{interval}':
    summary: Search Prices from ENTSO-E by a daily interval ahead of back in time.
    description: >-
      By providing an area code and interval, you can look up prices from ENTSO-E in the specified area
      by a daily interval, just by specifying the interval as a number. This could be 50 days or -50
      days from todays date.
    get:
      tags:
        - PriceIntervalController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByDailyInterval
      summary: Get ENTSO-E Market Prices by a daily interval.
      description: |-
        ## Interval by intervalType: Day
        __Interval: 12 days ahead from todays date__
        * /prices/FR/by-interval/day/12
        
        __Interval: 50 days back from todays date__
        * /prices/FR/by-interval/day/-50
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: interval
        description: >-
          The interval/amount of days you want to look up in the past or future, e.g. 7 for 7 days
          ahead, or -7 for 7 days back in time.
          
          
          Max interval is 365 days (1 year).
        schema:
          type: integer
        in: path
        required: true
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
  '/prices/{areaCode}/by-interval/month/{interval}':
    summary: Search Prices from ENTSO-E by a monthly interval ahead of back in time.
    description: >-
      By providing an area code and interval, you can look up prices from ENTSO-E in the specified area
      by a monthly interval, just by specifying the interval as a number. This could be 5 months or -11
      months from todays date.
    get:
      tags:
        - PriceIntervalController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByMonthlyInterval
      summary: Get ENTSO-E Market Prices by a monthly interval.
      description: |-
        ## Interval by intervalType: Month
        __Interval: 5 months ahead from todays date__
        * /prices/FR/by-interval/month/5
        
        __Interval: 5 months back from todays date__
        * /prices/FR/by-interval/day/-5
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: interval
        description: >-
          The interval/amount of months you want to look up in the past or future, e.g. 1 for 1
          month ahead, or -5 for 5 months back in time.
          
          
          Max interval is 12 months (1 year).
        schema:
          type: integer
        in: path
        required: true
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
  '/prices/{areaCode}/by-interval/week/{interval}':
    summary: Search Prices from ENTSO-E by a weekly interval ahead of back in time.
    description: >-
      By providing an area code and interval, you can look up prices from ENTSO-E in the specified area
      by a weekly interval, just by specifying the interval as a number. This could be 4 weeks or -10
      weeks from todays date.
    get:
      tags:
        - PriceIntervalController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByWeeklyInterval
      summary: Get ENTSO-E Market Prices by a weekly interval.
      description: |-
        ## Interval by intervalType: Week
        __Interval: 10 weeks ahead from todays date__
        * /prices/FR/by-interval/week/10
        
        __Interval: 20 weeks back from todays date__
        * /prices/FR/by-interval/day/-20
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: interval
        description: >-
          The interval/amount of weeks you want to look up in the past or future, e.g. 7 for 7 weeks
          ahead, or -7 for 7 weeks back in time.
          
          
          Max interval is 53 weeks (1 year).
        schema:
          type: integer
        in: path
        required: true
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
  '/prices/{areaCode}/by-interval/year/{interval}':
    summary: Search Prices from ENTSO-E by a yearly interval ahead of back in time.
    description: >-
      By providing an area code and interval, you can look up prices from ENTSO-E in the specified area
      by a yearly interval, just by specifying the interval as a number. This is limited to 1 year ahead
      or -1 for 1 year in the past.
    get:
      tags:
        - PriceIntervalController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByYearlyInterval
      summary: Get ENTSO-E Market Prices by a yearly interval.
      description: |-
        ## Interval by intervalType: Year
        __Interval: 1 year ahead from todays date__
        * /prices/FR/by-interval/year/1
        
        __Interval: 1 year back from todays date__
        * /prices/FR/by-interval/day/-1
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: interval
        description: >-
          The interval/amount of days you want to look up in the past or future. This can either be
          1 or -1 due to the limit by ENTSO-E.
          
          Max interval is 1 year.
        schema:
          type: integer
        in: path
        required: true
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
...
  '/prices/{areaCode}/by-range/{fromDateTime}':
    summary: Search Prices from ENTSO-E between todays date and the specified from date/time.
    description: >-
      By providing an from date, in the format of either YYYYMMDD (20221031) or YYMMDDHHMM
      (202210312300), you can look up prices from ENTSO-E in the specified area, between the "from-date"
      specified and todays date. 
      
      
      Max 365 days (1 year) from todays date. date.
    get:
      tags:
        - PriceRangeController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByFromDate
      summary: Get ENTSO-E Market Prices between a specified date and todays date.
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: fromDateTime
        description: >-
          The date/time that you want the prices from/to. This could be a date/time in the past, or
          the future, as long as it is no more than 365 days (1 year).
          
          Has to be specified in the format of either YYYYMMDD (20221031) or YYMMDDHHMM
          (202210312300).
        schema:
          type: string
        in: path
        required: true
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
  '/prices/{areaCode}/by-range/{fromDateTime}/{toDateTime}':
    summary: Search Prices from ENTSO-E between 2 specified dates.
    description: >-
      By providing a from and to date, in the format of either YYYYMMDD (20221031) or YYMMDDHHMM
      (202210312300), you can look up prices from ENTSO-E in the specified area, between the 2 specified
      dates.
      
      
      Max 365 days (1 year) between the 2 dates.
    get:
      tags:
        - PriceRangeController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByFromToDate
      summary: Get ENTSO-E Market Prices between a 2 dates.
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: fromDateTime
        description: >-
          The date/time that you want the prices from. This can be no more than 365 days (1 year)
          apart from the "to"-date.
          
          
          Has to be specified in the format of either YYYYMMDD (20221031) or YYMMDDHHMM
          (202210312300).
        schema:
          type: string
        in: path
        required: true
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
      -
        name: toDateTime
        description: >-
          The date/time that you want the prices to. This can be no more than 365 days (1 year)
          apart from the "from"-date.
          
          
          Has to be specified in the format of either YYYYMMDD (20221031) or YYMMDDHHMM
          (202210312300).
        schema:
          type: string
        in: path
        required: true
components:
  schemas:
    ErrorMessage:
      description: Error Message
      type: object
      properties:
        message:
          description: Error Message
          type: string
        details:
          description: ''
          type: array
          items:
            $ref: '#/components/schemas/ErrorMessageDetail'
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "message", "details" })
    ErrorMessageDetail:
      description: ''
      type: object
      properties:
        field:
          description: Field name
          type: string
        message:
          description: Information regarding the field error.
          type: string
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "field", "message" })
...
    AreaCode:
      description: |-
        Area Codes.
        * FR = France _(10YFR-RTE------C)_
        * DK1 = Denmark - West _(10YDK-1--------W)_
        * DK2 = Denmark - East _(10YDK-2--------M)_
        * DE = Germany _(10Y1001A1001A82H)_
        * SE1 = Sweden 1 _(10Y1001A1001A44P)_
        * SE2 = Sweden 2 _(10Y1001A1001A45N)_
        * SE3 = Sweden 3 _(10Y1001A1001A46L)_
        * SE4 = Sweden 4 _(10Y1001A1001A47J)_
        * NO1 = Norway 1 _(10YNO-1--------2)_
        * NO2 = Norway 2 _(10YNO-2--------T)_
        * NO3 = Norway 3 _(10YNO-3--------J)_
        * NO4 = Norway 4 _(10YNO-4--------9)_
        * NO5 = Norway 5 _(10Y1001A1001A48H)_
      enum:
        - FR
        - DK1
        - DK2
        - DE
        - SE1
        - SE2
        - SW1
        - SW2
        - SW3
        - SW4
        - NO1
        - NO2
        - NO3
        - NO4
        - NO5
        - FI
      type: string
    PricePoint:
      description: ''
      type: object
      properties:
        price:
          format: double
          description: ''
          type: number
        hour:
          description: ''
          type: string
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "hour", "price" })
    TimeInterval:
      description: ''
      required: []
      type: object
      properties:
        start:
          description: Interval Start Date/Time
          type: string
          example: '2022-10-15T22:00Z'
        end:
          description: Interval End Date/Time
          type: string
          example: '2022-10-17T22:00Z'
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "start", "end" })
    PricesResponse:
      description: ''
      required: []
      type: object
      properties:
        createdDateTime:
          description: Timestamp for Price Request creation.
          type: string
          example: '2022-10-19T05:00:55.466364'
        area:
          description: Area Name
          type: string
          example: FR - France
        areaCode:
          description: The ENTSO-E Area Code
          type: string
          example: 10YFR-RTE------C
        timeInterval:
          $ref: '#/components/schemas/TimeInterval'
          description: The Interval of the prices requested
        intervalDays:
          description: A list of the daily price information.
          type: array
          items:
            $ref: '#/components/schemas/IntervalDay'
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "area", "areaCode", "timeInterval",
        "intervalDays", "createdDateTime" })
    IntervalDay:
      description: The IntervalDay represents 1 day in the interval/range of the prices
      type: object
      properties:
        currency:
          description: Price Currency
          type: string
          example: EUR
        measureUnit:
          description: 'The measurement Unit related to the price, e.g. MWH.'
          type: string
          example: MWH
        hourPrices:
          description: A list of hourly prices throughout this Interval Date.
          maxItems: 24
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/PricePoint'
        currentDateInterval:
          $ref: '#/components/schemas/TimeInterval'
          description: The Interval of the current date
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "currentDateInterval", "currency",
        "measureUnit", "hourPrices" })
  responses:
    400BadRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          examples:
            Invalid Query Paramter:
              value:
                message: 'Invalid Query Parameter: interval=test'
                details:
                  -
                    field: interval
                    message: >-
                      Field should contain a value of type 'Integer'. Examples: 0, 1,
                      123, -1, -20 etc.
            Invalid Month Interval:
              value:
                message: Invalid parameters provided. See Details for more information.
                details:
                  -
                    field: interval
                    message: >-
                      Monthly Intervals are limited to +/- 12 months (1 year). If you
                      want to search for months further back in time, please use from/to
                      or year/month parameters instead, e.g.
                      /prices/FR?year=2020&month=10, or
                      /prices/FR?from=20201001&to=20201231
      description: >-
        When receiving a Bad Request Error Response, it means that the input provided in the API call
        is invalid.
        
        That could be an invalid combination of query paramters, an invalid ENTSO-E Security Token or
        invalid input in the API Request paramteres.
    500InternalServerError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          examples:
            Internal Error:
              value:
                message: An error occurred and has been reported to developers.
      description: >-
        Error Message Response when an Internal Server error happened, e.g. ENTSO-E does not respond
        etc.
        
        This error is not related to the request input, which means that the request is valid.
...